Change API as identified on devel mailing list.
Here is the email:
After some investigation into an error reported by valgrind, I'm proposing an
API change for the MoveCopyMessages() function.

The signature is currently
enum MAPISTATUS MoveCopyMessages(mapi_object_t *obj_src,  mapi_object_t 
*obj_dst,  mapi_id_t *message_id,  bool WantCopy)

The problem is the message_id array. The subtle part is that it needs to be 
terminated with a null mapi_id_t*, because of this:
for (request.count = 0; message_id[request.count]; request.count++);

That is a bit error prone - enough so to be wrong in the mapitest.

Now we could just fix the mapitest and document the requirements, or we could 
change the signature.

Two ideas for a different signature:
1. We could add a uint16_t count to the signature, that says how long the 
array is.
2. We could use a better data structure than mapi_id_t*. I'm suggesting
mapi_id_array_t:
enum MAPISTATUS MoveCopyMessages(mapi_object_t *obj_src,
                                          mapi_object_t *obj_dst,
                                          mapi_id_array_t *message_id,
                                          bool WantCopy)