Patch From Erik Hovland <erik@hovland.org>
In torture_rpc_mapi_sendattach a call is made to the system function
read at line 199. The return value of read is assigned to read_size,
which happens to be of unsigned type. So if read returns -1 (which it
can) it will not be noticed. The patch changes the type to ssize_t
(signed) and checks for -1 as well as zero to break out of the loop.

Ref ticket #156