Apply patch from Erik Hovland, from ticket #162.
From the original patch:
"The problem is that when you do stat first w/ a file string anyone can
exploit the time in between to do whatever they want with the file that
the string is a name for. Instead, try to open straight away. Then fstat
w/ the file descriptor."

I think the patch is fine, although I didn't run the torture test.