- Push Brad's CreateFolder implementation after turning hardcoded fields to generic one.
Note: current implementation is incorrect as it states that any folder is referenced
within openchange.ldb while only system/top folders are. We have a mapistore URI for
system/top folders, not for subfolders which are stored directly within specified
mapistore backend.

However the implementation is almost correct when it comes to create folders directly
under / in the user mailbox.