Send a flagged PropertyRow when GetProps is called with mapistore Add a nasty hack and call this fake MAPI_E_NOT_FOUND property value for GetProps if the object doesn't exist.
In openchangeclient we are calling get_container_class using the parent object
(Mailbox root folder) and the FID of the 2-level down folder. This causes
openchange server to crash. The hack exposed above will help temporary workaround
this problem while we figure out/remember the real problem: Why do we have
to call the parent to retrieve the container class?