Fix a bug in session management API:
The mpm_session API uses id, id2, node and context_id from the dcerpc
session to uniquely identify a session context. 

However when Outlook calls EcDoConnect multiple times, it still uses
the same dcerpc session. With previous API we were freeing the session
while it was still in used - which lead in unstable and unpredictable
crashes.

The following patch adds a ref_count to the mpm_session API and server
code has been changed for emsmdb and nspi server to increment the
ref_count when the session already exists or create a new session if
none is already stored.

Also the mpm_session code checks for the ref_count value and ensures
it is 0 before calling the destructor.