Minor fixes for property dumping in mapidump
Add support for PT_OBJECT and PT_CLSID property types.

Fix valgrind error in PT_UNICODE / PT_STRING8. The problem was that if
the string was empty (0x0000 for unicode), when we did a cast to a
uint32_t to test if it was MAPI_E_NOT_FOUND, we're using invalid memory.
The fix is good for unicode - perhaps could be better for string8.