Fix CopyTo test case.
The problem was caused by use of PR_CONVERSATION_TOPIC as a property,
which is read-only for the client after the first save in Exchange 2010.
So we use PR_SUBJECT instead, removing this property on the target message
to allow testing of the "no-overwrite" case.

Also, we can't use CopyFlagsNoOverwrite|CopyFlagsMove on Exchange 2010
(which was fixed in r2147 for another case).

I also changed the filename for the attachments to be short enough to fit
into the "DOS" convention, to avoid server-side truncation.

That now passes, and resolves #227.