Update NSPI tests to report errors properly.
This increases the amount of failing NSPI tests by two (i.e. get reported as
broken now) - NSPI-MODLINKATT and NSPI-MODPROPS.

Also fix one test (NspiSeekEntries), which was a bit flaky (it worked in some
configurations, and not in others). The problem turned out to be that we were
using the profile->username for PR_DISPLAY_NAME value. That doesn't always
hold, so we now use the EMSMDB login szDisplayName instead.