Fix NSPI ModProps test.
It appears that the PR_DISPLAY_NAME is not changeable, which is probably
a good thing because the test didn't try to set it back, and a random
PR_DISPLAY_NAME would almost certainly have confused things.

The new test uses PR_OFFICE_LOCATION, and resets it at the end.