Commit an initial (incomplete) async notification implementation.
This is only client side. It comes with a simple test application
that basically just blocks waiting for the async notification.

This isn't the final API, and there are some structural issues
with the way in which Subscribe() works. 

Also, works a lot better on Exchange 2007, still need to work on
the Exchange 2010 issues.