mapistore: rename get_fid_by_name to be get_uri_by_name in backend ops
This reflects actual usage / arguments. It might be possible to avoid
doing any backend operation for this (just use indexing DB instead), but
this allows us to check if the backend folder has been created by some
other process and would create a problem at next update. Not sure if that
is important yet.

Fix implementation of this backend function in FSOCPF backend.

Add unit testing of duplicate folder creation, which exercises this.