Patch from Sean Finney: Okay, here's something that gets pretty darned close, probably "close enough" for the next release. With this patch (along with the previous, already committed patches), one can create a profile that uses kerberos (and even the existing kerberose credentials cache for SSO users). to create such a profile:
mapiprofile -c -P kerbtest --username <username> --nopass -k yes --address <exchangeserver>

afterwards, if you run openchangeclient -m -p kerbtest, you will see the
fetched kerberos tickets in your klist output. Ideally this could be
completely automagic and not require a -k flag at all (nor the username), but
the fix for that would be a bit more complicated and is left for a future
exercise. In the meantime this solution will not get in the way and isn't too 
bad, at least IMHO :)

Addresses ticket #336.

Thanks for the patch!