Commit patch for issue #385:
Akhil found an issue [1] in OpenChange's code, that dcerpc_EcRRegisterPushNotification_r 
inside emsmdb_register_notification can succeed, but the retval can still be false. 
In this case the emsmdb_register_notification returns success, but it should return failure, 
because notifications aren't available. Evolution-mapi depends on this result code and Subscribe 
to notifications only if this succeeded.