openchangedb_property: Search String properties by PROP ID
This way we handle both PT_UNICODE and PT_STRING8 type of properties.

This commit fixes #7469 - can't share calendar bug
The problem is that Outlook requests for PidTagContainerClass_string8
property on Calendar folder. We assume this property should be UNICODE
only and hence we return NOT_FOUND. This confuses Outlook a lot and
it refuses to open a shared calendar.

Signed-off-by: Kamen Mazdrashki <kmazdrashki@zentyal.com>