fix regression on openchangedb ldb backend related with public folders
Folder ids for public folders must be obtained in a special way when
using mysql backend, but if we are using ldb backend we have to obtain
them like always (using mapistore_indexing_*).

openchangedb_is_public_folder_id will return always false when we
are using ldb backend and will return the proper value (depending of the
folder id given as parameter) with mysql backend.