mysql: cache of connections
It's not a connection pool because as we don't have threads the connection can
be used from each component and the component that ask for a connection doesn't
need to return it to the pool. This is just a dict like:

         { connection_string: connection * }

The common case, where we use only 1 mysql server, will be to have one
connection used everywhere.