Enable async notifications.
To get notifications, we must first inform the server which ones we're interested in.  Subscribe() expects notify_ctx to be available, so we initialize it immediately after establishing an async connection.

When we do wait, the spec says the server should complete the ROP after 5minutes, but in practice this doesn't seem to be a hard limit.  If the server takes too long, then we'll timeout.  In that case we just issue a new request and continue waiting.