Avoid read over boundaries error in get_AppointmentRecurrencePattern
To avoid this error, get_AppointmentRecurrencePattern now ignores reserved block length fields if not present in the pattern.
This should avoid crashes with clients that send the message without caring about this fields.
However if they are exceptions to the recurrence we cannot do so because the ExtendedException array is just after the ReservedBlock1. In that case we continue to return error.