mapiproxy: emsmdb/nspi single samdb connection
In both emsmdp_context and emsabp_context there is a pointer to samdb
(which is ldb_context). This ldb connection is used read only and there
is no need to create a new one for each bound session.

There was a problem when samdb_url was set (because openchange run in
a samba member node) and too many open files error happening due to
idle sessions (remember we were openning one ldap connection per emsdbp_context).