new CHANGES.md
The purpose of this file is that openchange users can understand what
is being done without reading git log. This means the messages should
be user oriented and not developer oriented (like commit messages are).

There should be subcategories to be able to read it faster.

The categories could be things like: Fixed, Added, Changed,
Removed, Security, Performance, etc...

Inspired by https://github.com/olivierlacan/keep-a-changelog