oxcfxics: Give support to retrieve sharing props on demand
Check the PidTagMessageClass when looping over a table and
check if it is a Sharing object using PidTagMessageClass equals
to "IPM.Sharing" and retrieve all that sharing properties (around 30)
for that specific message as well as common table ones.

This method could be extensible to have different messages inside
the same message table.

This gives support to Sharing Calendar Invitation in Outlook 2007
from invitations sent by Outlook 2007 or greater