Create XML Sharing message for Request + Invitation requests
This has required to do a major refactoring to include
TargetRecipients attribute and Providers node without regenerated
one after the other.

If you deny a request with an invitation in the same message
the returned message includes as PidNameXSharingFlavor the value
of 5100 which is not in any specification but it is also managed
here.

Introduce get_emsmdbp_data to safely obtain a data_pointer
after getting properties from mapistore layer.

Constants about PidLidSharingFlavor values are stored in
mapistore.h as they can be used by mapistore backends