nspi: QueryRows now can accept a NumPos bigger or equal to table size.
This is necessary because Outlook 2010 sometime send a NSPI QueryRow where
the initial position is after the end of the table. For example,
using the search in the GAL.
If I interpret the requested delta as going backwards, the client behavior
is fine so I suspect in a sign error in the client side.