[nspi] Do not raise invalid bookmark error in addresses container without purported search
The "All address list" container is meant to be only used as organizer and does not contain
the purportedSearch attribute.
This case was not taken into account so it always gave a 'Invalid bookmark' error instead of
a empty list. This behaviour can be checked in Outlook selecting 'All address list' in the dropdown at
the 'Select names' dialog.

To take it into account the functions dcesrv_do_NspiQueryRows, emsabp_ab_container_enum and emsabp_ab_fetch_filter
has been modified.