Fix log level on OC_DEBUG wrapper
With this change, the log level set in samba configuration
is honored in OC_DEBUG wrapper. Before this patch, if a
developer sets the log level to 5, it requires a level 6
in samba configuration to be logged.

I think this should be rollbacked once OpenChange
uses its own debugging system.