Refactor provision to use current_fid when it is set
This fixes a regression introduced by 63f2136a when
we want to reuse a folder as special folder if one
already exists on Logon.

Without this patch, the client gets messed up because
the fid set is not corresponding to the right folder,
making appear MAPI_E_NOT_FOUND opening folders and
incongruent information in oxcfxics protocol when
hierarchy was synched.

An example was the creation the Calendar, Contacts
and Tasks special folders by the client ignoring the
one we create on first Logon.