oxcfxics: Set sort in changed messages before and conditionally
According to [MS-OXCFXICS] Section 3.2.5.9.1.1 messages must be
sorted only when OrderByDeliveryTime synchronisation extra flag is set.

We are setting the sort order before getting the count to
avoid mapistore backends to perform the same query twice. One
to get the table row count and another one to sort them. SetSortOrder
operation in mapistore backend may perform no operation until
a first row is fetched or the count is asked.