oxcfxics: Optimise message changes discard already in cnset_seen
Perform the search for message identifiers along with the change
number to discard a change whose is already included in the cnset_seen
before opening the message and get all its available properties.

The use case is when you are in the middle of synchronisation of
a business size mailbox with a complex cnset_seen meta property,
current implementation fetches all messages, thus the number
of already included changes is large. With this,
the discard of already included changes is faster and, therefore,
in Outlook 2013 at least, the timeout between two FastTransferGetBuffer
replies is not triggered and infinite loop with no full sync
does not happen.