mapitest: Give support to skip test when is flagged as NotInOpenChange
Some operations are not implemented in OpenChange server so we
can flag some of them as not implemented and do not fail the test suite
as a whole. The tests may be marked as skipped.

The current check to know if the server is OpenChange is quite a "hack"
and it requires a Logon ROP per test so it is not ideal. It would be
great to have a way to distinguish OpenChange server in EcDoConnect operation.