oxcfxics: Always include valid mid in IdsetGiven
Without it, the messages that are valid in the server that
the client must have the latest version are not included
in the initial IdSetGiven are returned back in the meta property.

This avoids sync issues in INBOX folder, which stops syncing, after
performing several movements at once. Although the bug is not
easy reproducible by hand, using automated tests, the bug was 100%
reproducible and with this commit is gone.

PS: 5 weeks of several engineers to fix this.