Check return in oxcfxics_push_messageChange
This check avoids a crash in the server when the Source Key can't be
retrieved, leaving the binary structure uninitialised.

This behaviour has been seen when there are MIDs in the database with a Replica ID that cannot
exist. This patch doesn't fix that bug, it just works the crash around.