ocdb: Return MAPI_E_NOT_FOUND only if fid is not in DB
For set_mapistore_URI and set_system_idx procedures.

mysql_affected_rows is only reporting the affected rows and
when the value has the desired value the number of affected
rows is 0. We can change this behaviour by setting CLIENT_FOUND_ROWS
flag on mysql_real_connect but I'd prefer to have the MySQL
connection as standard as possible and this case is not so often.

More details on the following link:

https://dev.mysql.com/doc/refman/5.5/en/mysql-affected-rows.html