Tentative fix for CR-1136.
The analysis of the message identifiers involved in every crash
tends to highlight that the mid is initially garbage pushed to
the database and the fequency of this crash is statistically very
low ]1/5000;1/500000[. This patch therefore initializes messageID
to 0 beforehand and add memory checks to prevent data corruption.

At this stage of the investigation, this is the only plausible
explanation available. If it does fix the issue, then I expect
this fix to help narrowing it down even further.