util-memcached: shared or not shared connections
shared connections will be cached and not shared connections will not.

Shared connection will be used from openchange-samba side (no threading
so it's safe to reused them). Right now used in (1) indexing-mysql backend
and in (2) mapistore notification.

Not shared connections will be used from dovecot openchange plugin.
The initialization of mapistore_notification context will have a flag
(set on loadparm_context) "mapistore:threading" as true that will be
used to create a memcached connection with shared flag disabled.
That is because this plugin has concurrency and we cannot use the same
memcached object between threads.