ocsmanager-oof: fix unicode issues
Both external and internal reply messages were being bad encoded (set oof)
and bad decoded (get oof config). This will fix exceptions on oof methods
were the user was trying to set a message with non-ascii characters.

When writting to file we need to use str but only try to encode if the
variable is unicode ('ñ'.encode('utf8') is bad, u'ñ'.encode('utf8') is ok).
When reading and returning the data as xml we need to transform the
output of b64decode (str) into unicode using decode method, because
the xml parser will try to convert it to str (using encode).