oxcfxics: Only return state of involved objects on Upload changes
When asking for GetTransferState after performing Upload changes
(additions, deletions or modifications) we are returning back the
change numbers and idsets in meta properties which are related with
those changes.

We have implemented by storing in the synccontext object an idset
with the involved objects (folder or message identifiers) on every
upload operation. Furthermore, we are storing the change number
to query to mapistore for changes greater or equal to that value
so that the modified patch is not that big.

With this change, we avoid race conditions where missing
new objects (mails, appointments) is possible as we were returning
current state including new ids and its associated change numbers
and the client will never ask for them again.