oxcfxics: Mimetise Online mode for SetReadFlag on offline mode
Basically, we set the read flag and then, we save the message.

Apart from that, we are checking the mapistore functions' return
values and logging the errors to an upper level.

This change allows the mapistore backends to commit the message
changes only when save changes message is called, likewise, online
mode works.