oxcfxics: Do not merge non-related CNs on getting state
By setting the single attribute at the beginning, the change numbers
are added to the rawidset without merging values in between that
could appear if new change numbers are generated during the uploading
of changes non-related with the identifiers involved in upload operations.

This usually happens when you mark several messages as read, sogo-openchange
backend updates the folder CN every operation on message is done including
those CNs in the returned `CNSetSeen` that could lead to inconsistencies
and creating sync issues.