libmapi: Merge only idset with valid ranges
Provided that the first operation for a folder is an upload
and there is no state uploaded by the client, the stored state
by the server is the initialization one, emsmdbp_object_synccontext_init,
is an idset with an invalid range.

Given that invalid idset, when getting the transfer state after
uploading operations, the returned `CNSetSeen` includes an invalid GLOBSET
range which could make the MAPI client to have sync issues and stop
syncing that folder.