ndr: Removed charset attribute
The exchange.idl defined almost all string8 variables as having
'DOS' charset. The generated code for this charset raises errors
with non-ascii characters  values.

Since we do not know beforehand which encoding or code page is used
by the string8 variables, we remove the charset attribute.

This will remove all the charset checks and change the base type
of string pointers from char to uint8_t.

The failure when creating accounts with no-latin display names was
triggered by this error.