libmapi: Decoding GLOBSETs properly which have bitmask command
Two bugs were spotted:

* The *StartingValue* was set to a wrong value because the buffer_position
  was altered before stored additional data that is combined in memory
* The GLOBSET ranges were incorrectly set because the calculated
  value was based on the bit value, which checks the mask against, instead
  of the position in the loop.

Two unit tests were added that cover the 100% of the
`GLOBSET_parser_do_bitmask` code.

The consequence of this bug is described as follows:

Once an idset is incorrectly decoded for one of the fxics meta-properties,
when they are returned as the final result of an upload/download operation
in the FastTransfer Buffer, it returns an inconsistent idset with incorrect
GLOBSET ranges. For instance, having a folder whose identifiers the client
does not have leads to the client to stop syncing that folder.

For more details about *bitmask* command, check [MS-OXCFXICS]
Section 3.1.5.4.3.1.3.