oxcfxics: Perform CN restrictions on downloading changes to client
Using the given CNSet_Seen meta-property sent by the client we can ask
to the mapistore layer to retrieve only the missing messages.

We assume, likewise it was previously, the elements up to the first
globset range are not required as this element is the first CN seen by the client
upon first sync with empty state.