libmapi: Add PidTagRoamingBinary tag
As referenced by MSDN docs at the following URL:

https://msdn.microsoft.com/en-us/library/hh204511.aspx

This property contains the data stream associated with an IPM.Configuration message class message.
For instance, a message of class type IPM.Configuration.Autocomplete would be formatted as an Autocomplete Stream.

This adds as well a new way to add additional properties
not available in MS-OXPROPS available PDF.

In a nutshell, autocomplete in Outlook 2010 onwards is persistent
to profiles and closing/opening Outlook process.