Changed :days value on ocsmanager
Change :days value on OOF configuration to 1 (0 is not RFC-compliant)

https://tools.ietf.org/html/rfc5230#section-4.1

The ":days" argument is used to specify the period in which addresses
are kept and are not responded to, and is always specified in days.
The minimum value used for this parameter is normally 1.  Sites MAY
define a different minimum value as long as the minimum is greater
than 0