emsmdb: Distinguish authenticated user and logged user
The former is related with EcDoConnect and EcDoConnectEx RPC operations
and the user requires to be verified by an external authority in this
case Samba AD.

The latter is the user provided with Logon ROP and it is used to access
the user's mailbox to perform later other operations with that.

There were a mixture of them in a single variable emsmdbp_ctx->username
which led to wrong results and access to data from authenticated user
instead of the logged user and vice versa.